refactor(core): split color families into manual, identity, and binned modules
Extract resolveManualColorScale, resolveIdentityColorScale, and
resolveBinnedColorScale from the scale-color-families grab-bag so each
family owns its file (matching sequential/ordinal). Shared NA/unknown
helpers live in scale-color-family-helpers; uniqueValues stays private
to manual. scale-color.ts imports the three modules directly — no
re-export barrel.

Co-locate the 457-line pipeline-color-families test monolith under
tests/pipeline-color-scales/{sequential,binned,manual,identity,ordinal}
with shared fixtures.
