+++ .changeset/core-scale-color-families-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/core": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+refactor: split non-position color families into manual, identity, and binned modules
